refactor: extract receiver-threads from adas-manager
